ideater/modules/architecture_bot/requirements.txt
# Architecture Bot module requirements
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0